added logic using use params to be able to edit each golf club and redesigned the slider bar at the home page. This app is fully functional now

+++ client/src/utils/mutations.js
@@ -83,7 +83,7 @@ export const REMOVE_CLUB = gql`
 `
 
 export const ADD_DISTANCE = gql`
-    mutation addDistance($_id: ID, $clubAverage: Int, $clubLow: Int, $clubHigh: Int, $dateTested: String, $clubName: String) {
+    mutation addDistance($_id: ID, $clubAverage: Float, $clubLow: Float, $clubHigh: Float, $dateTested: String, $clubName: String) {
         addDistance(_id: $_id, clubAverage: $clubAverage, clubLow: $clubLow, clubHigh: $clubHigh, dateTested: $dateTested, clubName: $clubName) {
             _id
             username
